+++ sdk-media/src/main/kotlin/com/pexip/sdk/media/SecureCheckCode.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2023 Pexip AS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.pexip.sdk.media
+
+/**
+ * A secure check code.
+ *
+ * It is mainly used to verify whether the direct media connection is secure by comparing both local
+ * and remote values.
+ *
+ * @property value the actual value of this [SecureCheckCode]
+ */
+@JvmInline
+public value class SecureCheckCode(public val value: String)