+++ CHANGELOG.md
@@ -13,6 +13,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   `InfinityService.RegistrationStep.requestBuilder`, `InfinityService.ParticipantStep.conferenceStep`,
   `InfinityService.CallStep.participantStep` that give access to previous steps
 - `Conference.referer` to improve call transfer flow
+- `MediaConnection.secureCheckCode` that can be used to verify whether a direct media connection is
+  secure
 
 ### Changed